finish factoring out message_serializers.py
