+++ nanome/api/callbacks.py
@@ -1,21 +1,9 @@
 import logging
-from ._hashes import Hashes
 
 from nanome.api.integration import Integration, IntegrationRequest
 
 
 logger = logging.getLogger(__name__)
-
-
-def integration(network, args, request_id):
-    integration = network._plugin.integration
-    request = IntegrationRequest(args[0], args[1], args[2], network)
-    name = Hashes.HashToIntegrationName[args[1]]
-    Integration._call(integration, name, request)
-
-
-def run(network, args, request_id):
-    network._on_run()
 
 
 def simple_callback_arg_unpack(network, arg, request_id):
