+++ nanome/api/integration/__init__.py
@@ -1,4 +1,4 @@
 from . import *
 from .integration_request import IntegrationRequest
 from .integration import Integration
-from . import messages
+from . import messages, callbacks
